Assembly and mechanical function test
On September 11th 2023 the wheel cases were assembled and the mechanical functions (movement, spring function) were tested. Additionally the robotic arm was placed on the middle plate to test it's structural stability. Overall our goals were reached: mechanical function worked perfectly and the middle plate had no issues supporting the robotic arm. It is assumed that any additional weight put on the middle plate (PC,  GPU stand) is insignificant and therefore won't change that assesment. 
However some problems occured: Assembly of the wheel cases was unpleasant since the mecanum wheels used for the plattform a different from the models initialy used in the CAD model. So with this design quick reasemmbly is not easy. Redesigning the wheel cases for the bigger mecanum wheels might be in order but the current design works.